Step 3: remove sleep(), add onetimerun check

+++ FlaskTemplate/preinit.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-# echo "import pty; pty.spawn('/bin/sh')" > /tmp/asdf.py
-# python /tmp/asdf.py
-# runAsRoot="
-curl https://packages.microsoft.com/keys/microsoft.asc | apt-key add -
-curl https://packages.microsoft.com/config/ubuntu/18.04/prod.list > /etc/apt/sources.list.d/mssql-release.list
-apt-get update
-ACCEPT_EULA=Y apt-get install -y msodbcsql17
-ACCEPT_EULA=Y apt-get install -y mssql-tools
-apt-get install -y build-essential
-echo 'export PATH="$PATH:/opt/mssql-tools/bin"' >> ~/.bashrc
-source ~/.bashrc
-source "antenv"/bin/activate
-pip install pyodbc
+if [ ! -f myfile ]
+then
+    curl https://packages.microsoft.com/keys/microsoft.asc | apt-key add -
+    curl https://packages.microsoft.com/config/ubuntu/18.04/prod.list > /etc/apt/sources.list.d/mssql-release.list
+    apt-get update
+    ACCEPT_EULA=Y apt-get install -y msodbcsql17
+    ACCEPT_EULA=Y apt-get install -y mssql-tools
+    apt-get install -y build-essential
+    echo 'export PATH="$PATH:/opt/mssql-tools/bin"' >> ~/.bashrc
+    source ~/.bashrc
+    source "antenv"/bin/activate
+    pip install pyodbc
+    touch myfile
+else
+    echo "Looks like all packages present. Nothing to do"
+fi
+
 # exec su root -s /bin/sh -c "$runAsRoot"
